refactor: map models in several files
